repo cleanup
move sucks stuff to its own folder
