+++ sucks/log_clean.py
