+++ sucks/setup.py
