+++ sucks/sucks.sh
