+++ sucks/sucks/cli.py
