+++ sucks/sucks/const.py
@@ -1,0 +1,8 @@
+import logging
+LOGGER = logging.getLogger(__name__)
+
+#ecovacs constants
+#init constants
+ECOVACS_DEVICES = "ecovacs_devices"
+DOMAIN = "ecovacs"
+CONF_CONTINENT = "continent"
