+++ sucks/tests/test_cli.py
