+++ sucks/tests/test_commands.py
